requirements.txt
fastmcp>=3.2.0
boto3>=1.34.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
